+++ src/event_loop.c
@@ -1509,37 +1509,20 @@ void event_loop_post(struct event_loop *event_loop, enum event_type type, void *
     sem_post(event_loop->semaphore);
 }
 
-bool event_loop_init(struct event_loop *event_loop)
+bool event_loop_begin(struct event_loop *event_loop)
 {
     if (!memory_pool_init(&event_loop->pool, KILOBYTES(512))) return false;
+
+    event_loop->semaphore = sem_open("yabai_event_loop_semaphore", O_CREAT, 0600, 0);
+    sem_unlink("yabai_event_loop_semaphore");
+    if (event_loop->semaphore == SEM_FAILED) return false;
 
     event_loop->head = memory_pool_push(&event_loop->pool, sizeof(struct event_loop_item));
     event_loop->head->next = NULL;
     event_loop->tail = event_loop->head;
 
-    event_loop->is_running = false;
-    event_loop->semaphore = sem_open("yabai_event_loop_semaphore", O_CREAT, 0600, 0);
-    sem_unlink("yabai_event_loop_semaphore");
-
-    return event_loop->semaphore != SEM_FAILED;
-}
-
-bool event_loop_begin(struct event_loop *event_loop)
-{
-    if (event_loop->is_running) return false;
-
     event_loop->is_running = true;
     pthread_create(&event_loop->thread, NULL, &event_loop_run, event_loop);
-
-    return true;
-}
-
-bool event_loop_end(struct event_loop *event_loop)
-{
-    if (!event_loop->is_running) return false;
-
-    event_loop->is_running = false;
-    pthread_join(event_loop->thread, NULL);
 
     return true;
 }