+++ src/event_loop.h
@@ -88,9 +88,7 @@ struct event_loop
     struct event_loop_item *tail;
 };
 
-bool event_loop_init(struct event_loop *event_loop);
 bool event_loop_begin(struct event_loop *event_loop);
-bool event_loop_end(struct event_loop *event_loop);
 void event_loop_post(struct event_loop *event_loop, enum event_type type, void *context, int param1, volatile uint32_t *info);
 
 #endif