+++ src/process_manager.h
@@ -31,8 +31,6 @@ struct process *process_create(ProcessSerialNumber psn);
 struct process *process_manager_find_process(struct process_manager *pm, ProcessSerialNumber *psn);
 void process_manager_remove_process(struct process_manager *pm, ProcessSerialNumber *psn);
 void process_manager_add_process(struct process_manager *pm, struct process *process);
-void process_manager_init(struct process_manager *pm);
 bool process_manager_begin(struct process_manager *pm);
-bool process_manager_end(struct process_manager *pm);
 
 #endif