+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change launchd service to only restart automatically upon crashes [#1755](https://github.com/koekeishiya/yabai/issues/1755)
 
 ## [5.0.5] - 2023-05-21
 ### Changed